Added extra behavior to the AI
Added forward looping of the AI and The ability to set the looping - on the Patrol point  connecting the the behavior tree.
Added navigating back throught the patrol points creating a deincrement BTT. This in turn made it so i had to add a direction modifier to the behavior tree.

Plans to create backward looping of the patrol points depending on the direction setting.
this will probably force me to modify how the Increment and deincrement BTT's handle fail states